requirements.txt
accelerate
diffusers
facexlib
gradio==5.23.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
insightface
numpy==1.26.4
onnxruntime==1.19.2
opencv-python
optimum-quanto
peft==0.15.0
pillow==10.4.0
pillow-avif-plugin==1.5.0
pillow-heif==0.21.0
sentencepiece==0.2.0
torch
torchvision
transformers
netifaces